requirements.txt
python=>3.0
django=>2.2
django-import-export
